LPS-53684 move methods into new util class
